refactor: reorganize broadcaster modules and fix circular imports
