--- bsv/broadcaster_core/__init__.py
@@ -1,2 +1,0 @@
-# This file makes bsv.broadcaster a package.
-from .broadcaster import Broadcaster, BroadcastResponse, BroadcastFailure
