+++ bsv/chaintrackers/whatsonchain.py
@@ -1,4 +1,5 @@
 from typing import Optional, Dict
+from typing import Any
 
 from bsv.chaintracker import ChainTracker
 from bsv.http_client import HttpClient, default_http_client
